간단한 게시판 API 구현

+++ src/main/java/com/portfolio/mvc/controller/BoardController.java
@@ -49,8 +49,9 @@ public class BoardController {
 	 * @param board
 	 */
 	@GetMapping("/save")
-	public int save(Board board) {
-		return boardService.save(board);
+	public int save(Board parameter) {
+		boardService.save(parameter);
+		return parameter.getBoardSeq();
 	}
 
 	/**
@@ -58,7 +59,12 @@ public class BoardController {
 	 * @param boardSeq
 	 */
 	@GetMapping("/delete/{boardSeq}")
-	public void delete(@PathVariable int boardSeq) {
+	public boolean delete(@PathVariable int boardSeq) {
+		Board board = boardService.get(boardSeq);
+		if (board == null) {
+			return false;
+		}
 		boardService.delete(boardSeq);
+		return true;
 	}
 }
